[ShapeInfer] Renamed sequnce_generator.hpp -> sequence_generator.hpp
