+++ src/common/transformations/src/transformations/common_optimizations/pull_through_reduce.cpp
@@ -15,7 +15,7 @@
 #include "openvino/op/util/reduction_base.hpp"
 #include "openvino/pass/pattern/op/or.hpp"
 #include "openvino/pass/pattern/op/wrap_type.hpp"
-#include "sequnce_generator.hpp"
+#include "sequence_generator.hpp"
 #include "transformations/utils/utils.hpp"
 
 namespace {
