+++ src/core/shape_inference/include/sequence_generator.hpp
