+++ src/core/tests/eval.cpp
@@ -65,7 +65,7 @@
 #include "openvino/op/transpose.hpp"
 #include "openvino/op/unsqueeze.hpp"
 #include "openvino/runtime/tensor.hpp"
-#include "sequnce_generator.hpp"
+#include "sequence_generator.hpp"
 #include "utils/eval_utils.hpp"
 
 using namespace std;
