+++ src/core/tests/type_prop/squeeze.cpp
@@ -11,7 +11,7 @@
 #include "openvino/op/gather.hpp"
 #include "openvino/op/shape_of.hpp"
 #include "openvino/op/unsqueeze.hpp"
-#include "sequnce_generator.hpp"
+#include "sequence_generator.hpp"
 
 using namespace std;
 using namespace ov;
