+++ src/core/tests/type_prop/variadic_split.cpp
@@ -9,7 +9,7 @@
 #include "openvino/core/validation_util.hpp"
 #include "openvino/op/broadcast.hpp"
 #include "openvino/op/shape_of.hpp"
-#include "sequnce_generator.hpp"
+#include "sequence_generator.hpp"
 
 using namespace std;
 using namespace ov;
